+++ lizard_map/utility.py
@@ -43,3 +43,14 @@ def analyze_http_user_agent(http_user_agent):
         device = 'iPad'
 
     return {'device': device}
+
+
+def get_host():
+    """Get the current host.
+
+    Needed in the multitancy Lizard 5 site for cache keys.
+    """
+    host = ''
+    if hasattr(request, 'get_host'):
+        host = request.get_host()
+    return host